requirement.txt
streamlit>=1.44.1
plotly>=6.0.1
numpy
opencv-python
matplotlib
torch
torchvision
Pillow